+++ packages/components/organisms/f-checkout/src/components/Checkout.vue
@@ -83,7 +83,7 @@
 <script>
 import { validationMixin } from 'vuelidate';
 import { required, email } from 'vuelidate/lib/validators';
-import { mapState, mapActions } from 'vuex';
+import { mapActions, mapState } from 'vuex';
 
 import Alert from '@justeat/f-alert';
 import '@justeat/f-alert/dist/f-alert.css';
@@ -111,7 +111,7 @@ import { CHECKOUT_METHOD_DELIVERY, TENANT_MAP, VALIDATIONS } from '../constants'
 import checkoutValidationsMixin from '../mixins/validations.mixin';
 import EventNames from '../event-names';
 import tenantConfigs from '../tenants';
-import mapUpdateCheckoutRequest from '../services/mapper';
+import { mapUpdateCheckoutRequest } from '../services/mapper';
 
 export default {
     name: 'VueCheckout',
@@ -315,6 +315,7 @@ export default {
 
     async mounted () {
         await this.initialise();
+        this.trackInitialLoad();
     },
 
     methods: {
@@ -329,6 +330,11 @@ export default {
             'updateCustomerDetails',
             'updateUserNote',
             'placeOrder'
+        ]),
+
+        ...mapActions('analytics', [
+            'trackInitialLoad',
+            'trackFormInteraction'
         ]),
 
         /**
@@ -565,11 +571,11 @@ export default {
             }
         },
 
+        /**
+         * Emit `CheckoutFailure` event with error data
+         * Update `genericErrorMessage` to display correct errorMessage for passed error
+         */
         handleErrorState (error) {
-            /*
-            * Emit `CheckoutFailure` event with error data
-            * Update `genericErrorMessage` to display correct errorMessage for passed error
-            */
             let thrownErrors = error;
 
             // Ideally we would use optional chaining but it doesn't currently work with Storybook
@@ -610,15 +616,19 @@ export default {
             }
         },
 
+        /**
+         * Check form is valid - no inline messages
+         * If form is valid try to call `submitCheckout`
+         * Catch and handle any errors
+         */
         async onFormSubmit () {
-            /*
-            * Check for is valid - no inline messages
-            * If form is valid try to call `submitCheckout`
-            * Catch and handle any errors
-            */
+            this.trackFormInteraction({ action: 'submit' });
+
             if (!this.isFormValid()) {
                 const validationState = validations.getFormValidationState(this.$v);
+
                 this.$emit(EventNames.CheckoutValidationError, validationState);
+                this.trackFormInteraction({ action: 'inline_error', error: validationState.invalidFields });
 
                 this.$logger.logWarn(
                     'Checkout Validation Error',
@@ -632,6 +642,7 @@ export default {
 
             try {
                 await this.submitCheckout();
+                this.trackFormInteraction({ action: 'success' });
             } catch (error) {
                 this.handleErrorState(error);
             } finally {
@@ -639,26 +650,26 @@ export default {
             }
         },
 
+        /**
+         * Check to see if any `Vuelidate` validation errors
+         */
         isFormValid () {
-            /*
-            * Check to see if any `Vuelidate` validation errors
-            */
             this.$v.$touch();
             return !this.$v.$invalid;
         },
 
-        /*
-        * Use phone validation in `f-services` to check if customer number is
-        * valid in current locale
-        */
+        /**
+         * Use phone validation in `f-services` to check if customer number is
+         * valid in current locale
+         */
         isValidPhoneNumber () {
             return validations.isValidPhoneNumber(this.customer.mobileNumber, this.$i18n.locale);
         },
 
-        /*
-        * Use postcode validation in `f-services` to check if customer postcode is
-        * valid in current locale
-        */
+        /**
+         * Use postcode validation in `f-services` to check if customer postcode is
+         * valid in current locale
+         */
         isValidPostcode () {
             return validations.isValidPostcode(this.address.postcode, this.$i18n.locale);
         },
@@ -734,7 +745,7 @@ export default {
     .c-spinner {
         margin: 0 auto;
     }
-  }
+}
 
 .c-checkout {
     padding-top: spacing(x6);